Curriculum/Module 05 - Hacking with Python/assignments/hacking-with-python/assignment-3-math-solver/requirements.txt
pwntools==4.11.0
sympy==1.12
